backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-core==2.14.5
openai==1.3.7
agno==0.1.0
python-multipart==0.0.6
